Move tests closer to code under test
